<fix>[sdk]: unify the package structure
Resolves: ZSV-12254
