+++ sdk/src/main/java/org/zstack/sdk/logserver/LogServerState.java
@@ -1,4 +1,4 @@
-package org.zstack.sdk;
+package org.zstack.sdk.logserver;
 
 public enum LogServerState {
 	Enabled,
