+++ sdk/src/main/java/org/zstack/sdk/logserver/QueryLogServerResult.java
@@ -1,4 +1,4 @@
-package org.zstack.sdk;
+package org.zstack.sdk.logserver;
